feat(core,ffi,unity)!: widen NodeId ABI u32 -> u64 (#26)
Breaking ABI change, full-chain in one commit (core + frame blob v14 + FFI
surface + regenerated C# bindings + Unity Runtime), no compat shim.

NodeId(u64) layout: index bits[31:0] (full slotmap width, the 4096-node
synthetic-id hard panic is now structurally impossible) + generation
bits[55:32] (24-bit, 16.7M reuses/slot vs old 4096; debug_assert guard) +
tag byte bits[63:56]. Real nodes tag=0; synthetic layers occupy tag
segments (sub-pages 1..=15, V/H thumb 16/17 migrated from bit 30/29,
TF edit 32..=34, TF text 35, shadow 36..=47, inline-img 232..=255).
INVALID = u64::MAX.

Silent-decode mines fixed along the sweep: scroll wheel thumb sentinel
decode (old bit-mask missed new tag bits), FFI hit_test tag strip,
parent_id i32 truncation (now i64), C# event SOA offsets (28B->32B),
boxed uint.MaxValue assertions. Frame blob VERSION 13->14 (id columns
4B->8B); pkg format untouched (pkg.bin carries no NodeId).

Gates: workspace cargo test 1775 green; PublicApi -warnaserror and
HeadlessTests dotnet builds clean; Tests.Core 33/33; release dll rebuilt
and copied into unity/package/Plugins.

+++ crates/core/examples/dump_page.rs
@@ -82,7 +82,7 @@ fn main() {
     println!("\n── 渲染节点 world_matrix vs layout_rect（遍历 frame.nodes，查 blob 不匹配）──");
     // build node_id(index) → layout_rect 索引（scene 侧）
     use std::collections::HashMap;
-    let mut lr: HashMap<u32, (f32, f32, f32, f32)> = HashMap::new();
+    let mut lr: HashMap<u64, (f32, f32, f32, f32)> = HashMap::new();
     for n in scene.nodes.values() {
         lr.insert(
             n.id.0,
@@ -98,7 +98,7 @@ fn main() {
     let mut checked = 0;
     let mut samples = Vec::new();
     let mut orphan = 0;
-    let mut matches: HashMap<u32, (f32, f32)> = HashMap::new();
+    let mut matches: HashMap<u64, (f32, f32)> = HashMap::new();
     let wt_len = scene.world_transforms.len();
     println!(
         "  [diag] world_transforms.len()={} nodes.capacity()={}",
@@ -106,7 +106,7 @@ fn main() {
         scene.nodes.capacity()
     );
     // 重复 node_id 检测：同一 node_id 出现多次 → 后者覆盖前者 world_matrix？
-    let mut id_count: HashMap<u32, usize> = HashMap::new();
+    let mut id_count: HashMap<u64, usize> = HashMap::new();
     let mut dup_ids = 0;
     for rn in &frame.nodes {
         let c = id_count.entry(rn.node_id).or_insert(0);
