+++ crates/core/src/stage.rs
@@ -62,7 +62,7 @@ pub struct Stage {
     /// 上帧每节点 (header_hash, payload_hash)（node_id 键）。跨 tick 持续，供
     /// build_render_nodes 比较定 ChangeLevel。transient 不进 pkg（Stage 字段非 Scene 字段）。
     /// reload/节点数变 → clear → 下帧全 dirty（无基线）。
-    pub prev_node_hashes: std::collections::HashMap<u32, (u64, u64)>,
+    pub prev_node_hashes: std::collections::HashMap<u64, (u64, u64)>,
     /// 核心字形 atlas（v1.6 自绘字体）。render build 期 ensure 字形 UV，
     /// FFI 拉 R8 脏页上传。Stage 持有（非 Scene——atlas 是渲染资源，生命周期跨 tick）。
     pub glyph_atlas: crate::text::atlas::GlyphAtlas,
